fix: create mobile header component
